Show a loading spinner between login and auth checks
In case the activity gets loaded and we are already logged in there
is a short perio where we still wait for the person fetch to finish
so we can see if the user has the right permissions.

Instead of showing the "not logged in" message show a loading spinner.