Switch to the es5 build of pdf.js
The current version no longer works with safari 12 and shows an error suggesting
to use the es5 version instead. So give it a try.

The error: "The browser/environment lacks native support for critical functionality used by the PDF.js library"

The patch for enabling pdf-js to show signatures didn't apply anymore so make the replace pattern
more generic to match both the es6 and es5 build.