rollup: switch to the babel output plugin
Instead of transforming the input we now transform the output. This
has the downside that we can't share the helpers and duplicate them per chunk,
but also makes sure we cover all code that ends up in the bundle.

Also move terser aft er babel, so it gets minified last.