Add a funtion for guessing the signature count of a PDF file
We require the signature count in the future to force users to
do a manual positioning when there are already signatures in the file
since we can't add a new page for them.

This uses a simple regex, matching parts of the signatures we use (PAdES).

In case the result is wrong, worst case the user has to do extra work by positioning
or the auto positioning will put it at the top of the first page, which might not
be ideal, but the document is still signed either way.