Force users to provide a manual position for already signed documents
In case we get a signed document pdf-as in the backend can't add a new
page because implementations like Adobe Reader don't allow adding pages
after a signature has been added when checking the integrity.

A newer pdf-as version will take this into account and in the "auto" mode
position the signature on the first page. Since this usually means that
content will be overlaid this isn't a good "auto" mode.

To avoid this we check if the document is already signed in the frontend,
and force the user to place the signature manually in that case.