tables: set tabulator options directly instead of via render()
It was using a non-state property, so the binding only happend
if something else changed. We could make it a proper lit property,
but since it's only used/set in one place just set it directly.

This fixes the "failed" table sometimes not showing up, or only
showing up delayed.