Port to new app-shell
The activity name is now in the app shell, so remove it.
The description is only visible on the welcome page (and as title attribute), so
remove that too.

Remove any spacing at the top of the activities, so the text aligns with the menu etc.

Move to unitless line-height, as the app shell expects that now.